refactor: modularize Tkinter GUI
