+++ gui/progress.py
@@ -1,0 +1,3 @@
+"""Helper utilities for progress updates."""
+
+# This module is reserved for future progress-related helpers.
